requirements.txt
# pip3 install -r requirements.txt

py-clob-client==0.20.0
py_order_utils==0.3.2

pandas
gspread
gspread-dataframe
sortedcontainers
black==24.4.2
eth-utils===4.1.1
poly_eip712_structs==0.0.1

pytest==8.2.2
# python-dotenv==0.19.2
requests==2.32.3
websockets==12.0
cryptography==42.0.8
asyncio
requests
google-auth
python-dotenv>=1.0.0
web3==6.0.0
eth-account>=0.9.0
setuptools>=65.5.0
prometheus_client==0.14.1
PyYAML==6.0

scipy
websocket-client
matplotlib
seaborn